Moves Platform and Murmur3_x86_32 to o.a.s.u.sketch and makes them package private
