+++ common/sketch/src/main/java/org/apache/spark/util/sketch/CountMinSketchImpl.java
@@ -22,10 +22,7 @@ import java.io.UnsupportedEncodingException;
 import java.util.Arrays;
 import java.util.Random;
 
-import org.apache.spark.util.sketch.unsafe.Platform;
-import org.apache.spark.util.sketch.unsafe.hash.Murmur3_x86_32;
-
-public class CountMinSketchImpl extends CountMinSketch {
+class CountMinSketchImpl extends CountMinSketch {
   public static final long PRIME_MODULUS = (1L << 31) - 1;
 
   private int depth;
